+++ platform/core/src/extensions/ExtensionManager.test.js
@@ -6,7 +6,7 @@ import log from './../log.js';
 jest.mock('./../log.js');
 
 describe('ExtensionManager.js', () => {
-  let extensionManager, commandsManager;
+  let extensionManager, commandsManager, servicesManager, appConfig;
 
   beforeEach(() => {
     commandsManager = {
@@ -14,7 +14,17 @@ describe('ExtensionManager.js', () => {
       getContext: jest.fn(),
       registerCommand: jest.fn(),
     };
-    extensionManager = new ExtensionManager({ commandsManager });
+    servicesManager = {
+      registerService: jest.fn(),
+    };
+    appConfig = {
+      testing: true,
+    };
+    extensionManager = new ExtensionManager({
+      servicesManager,
+      commandsManager,
+      appConfig,
+    });
     log.warn.mockClear();
     jest.clearAllMocks();
   });
@@ -37,6 +47,24 @@ describe('ExtensionManager.js', () => {
       // Assert
       expect(extensionManager.registerExtension.mock.calls.length).toBe(3);
     });
+
+    it('calls registerExtension() for each extension passing its configuration if tuple', () => {
+      const fakeConfiguration = { testing: true };
+      extensionManager.registerExtension = jest.fn();
+
+      // SUT
+      const fakeExtensions = [
+        { one: '1' },
+        [{ two: '2' }, fakeConfiguration],
+        { three: '3 ' },
+      ];
+      extensionManager.registerExtensions(fakeExtensions);
+
+      // Assert
+      expect(extensionManager.registerExtension.mock.calls[1][1]).toEqual(
+        fakeConfiguration
+      );
+    });
   });
 
   describe('registerExtension()', () => {
@@ -49,19 +77,19 @@ describe('ExtensionManager.js', () => {
       expect(fakeExtension.preRegistration.mock.calls.length).toBe(1);
     });
 
-    it('calls preRegistration() passing configuration along with servicesManager and commandsManager instances for extension', () => {
-      const configuration = { config: 'Some configuration' };
-      extensionManager._servicesManager = { services: { TestService: {} } };
+    it('calls preRegistration() passing dependencies and extension configuration to extension', () => {
+      const extensionConfiguration = { config: 'Some configuration' };
 
       // SUT
-      const fakeExtension = { one: '1', preRegistration: jest.fn() };
-      extensionManager.registerExtension(fakeExtension, configuration);
+      const extension = { one: '1', preRegistration: jest.fn() };
+      extensionManager.registerExtension(extension, extensionConfiguration);
 
       // Assert
-      expect(fakeExtension.preRegistration.mock.calls[0][0]).toEqual({
-        servicesManager: extensionManager._servicesManager,
-        commandsManager: extensionManager._commandsManager,
-        configuration,
+      expect(extension.preRegistration.mock.calls[0][0]).toEqual({
+        servicesManager,
+        commandsManager,
+        appConfig,
+        configuration: extensionConfiguration,
       });
     });
 
@@ -135,8 +163,8 @@ describe('ExtensionManager.js', () => {
       );
     });
 
-    it('successfully passes a servicesManager and commandsManager instances to each module', () => {
-      extensionManager._servicesManager = { services: { TestService: {} } };
+    it('successfully passes dependencies to each module along with extension configuration', () => {
+      const extensionConfiguration = { testing: true };
 
       const extension = {
         id: 'hello-world',
@@ -147,11 +175,17 @@ describe('ExtensionManager.js', () => {
         getCommandsModule: jest.fn(),
       };
 
-      extensionManager.registerExtension(extension);
+      extensionManager.registerExtension(extension, extensionConfiguration);
 
-      expect(extension.getViewportModule.mock.calls[0][0]).toEqual({
-        servicesManager: extensionManager._servicesManager,
-        commandsManager: extensionManager._commandsManager,
+      Object.keys(extension).forEach(module => {
+        if (typeof extension[module] === 'function') {
+          expect(extension[module].mock.calls[0][0]).toEqual({
+            servicesManager,
+            commandsManager,
+            appConfig,
+            configuration: extensionConfiguration,
+          });
+        }
       });
     });
 
@@ -192,7 +226,7 @@ describe('ExtensionManager.js', () => {
           return {
             definitions: {
               exampleDefinition: {
-                commandFn: () => {},
+                commandFn: () => { },
                 storeContexts: [],
                 options: {},
               },