+++ platform/core/src/services/UIDialogService/index.js
@@ -17,59 +17,60 @@
  * @property {Object} contentProps The dialog content props.
  * @property {boolean} [isDraggable=true] Controls if dialog content is draggable or not.
  * @property {boolean} [showOverlay=false] Controls dialog overlay.
+ * @property {boolean} [centralize=false] Center the dialog on the screen.
+ * @property {boolean} [preservePosition=true] Use last position instead of default.
  * @property {ElementPosition} defaultPosition Specifies the `x` and `y` that the dragged item should start at.
- * @property {ElementPosition} position If this property is present, the item becomes 'controlled' and is not responsive to user input.
  * @property {Function} onStart Called when dragging starts. If `false` is returned any handler, the action will cancel.
  * @property {Function} onStop Called when dragging stops.
  * @property {Function} onDrag Called while dragging.
  */
 
-const uiDialogServicePublicAPI = {
-  name: 'UIDialogService',
-  dismiss,
-  dismissAll,
-  create,
+const name = 'UIDialogService';
+
+const publicAPI = {
+  name,
+  dismiss: _dismiss,
+  dismissAll: _dismissAll,
+  create: _create,
   setServiceImplementation,
 };
 
-const uiDialogServiceImplementation = {
+const serviceImplementation = {
   _dismiss: () => console.warn('dismiss() NOT IMPLEMENTED'),
   _dismissAll: () => console.warn('dismissAll() NOT IMPLEMENTED'),
   _create: () => console.warn('create() NOT IMPLEMENTED'),
 };
 
-function createUIDialogService() {
-  return uiDialogServicePublicAPI;
-}
-
 /**
  * Show a new UI dialog;
  *
- * @param {DialogProps} props { id, content, contentProps, onStart, onDrag, onStop, isDraggable, showOverlay, defaultPosition, position }
+ * @param {DialogProps} props { id, content, contentProps, onStart, onDrag, onStop, centralize, isDraggable, showOverlay, preservePosition, defaultPosition }
  */
-function create({
+function _create({
   id,
   content,
   contentProps,
   onStart,
   onDrag,
   onStop,
+  centralize = false,
+  preservePosition = true,
   isDraggable = true,
   showOverlay = false,
   defaultPosition,
-  position,
 }) {
-  return uiDialogServiceImplementation._create({
+  return serviceImplementation._create({
     id,
     content,
     contentProps,
     onStart,
     onDrag,
     onStop,
+    centralize,
+    preservePosition,
     isDraggable,
     showOverlay,
     defaultPosition,
-    position,
   });
 }
 
@@ -78,8 +79,8 @@ function create({
  *
  * @returns void
  */
-function dismissAll() {
-  return uiDialogServiceImplementation._dismissAll();
+function _dismissAll() {
+  return serviceImplementation._dismissAll();
 }
 
 /**
@@ -87,8 +88,8 @@ function dismissAll() {
  *
  * @returns void
  */
-function dismiss({ id }) {
-  return uiDialogServiceImplementation._dismiss({ id });
+function _dismiss({ id }) {
+  return serviceImplementation._dismiss({ id });
 }
 
 /**
@@ -106,14 +107,19 @@ function setServiceImplementation({
   create: createImplementation,
 }) {
   if (dismissImplementation) {
-    uiDialogServiceImplementation._dismiss = dismissImplementation;
+    serviceImplementation._dismiss = dismissImplementation;
   }
   if (dismissAllImplementation) {
-    uiDialogServiceImplementation._dismissAll = dismissAllImplementation;
+    serviceImplementation._dismissAll = dismissAllImplementation;
   }
   if (createImplementation) {
-    uiDialogServiceImplementation._create = createImplementation;
+    serviceImplementation._create = createImplementation;
   }
 }
 
-export default createUIDialogService;
+export default {
+  name,
+  create: ({ configuration = {} }) => {
+    return publicAPI;
+  },
+};